File created and served
Not an overly safe way to produce the results as the file is maintained
on the server until the next patient overwrites the information with
their own. But file can be grabbed and all medical information is
written on the text document and is downloaded by the patient after
warning.