Fix a bug with export information.
When two patients export their medical information at the same time, there would be a conflict. One patient might override the other patient's exported file, and one patient might be able to download someone else's information. 

Usernames are preppended to the file names to resolve the conflicts, so a restriction is added to `UserCreationForm` that forbids usernames starting with a '.' character, since that's not allowed by the file system.